Add MetadataPersistence classes
Add support to persist metadata in two different ways.
One is to continue storing as extended attributes in each
object file itself. The second mode stores the object metadata
as a hidden json file. This new mode will be useful when
support for extended attributes is not available. (i.e., nfs or smb shares)